% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{appendRDSBundle}
\alias{appendRDSBundle}
\title{Append objects to a RDSBundle file. If the first objects size is less than
the size of the index table of the existing file, instead of overwriting the
index table it is left in place and the new objects are appended at EOF.
This can be verified by using getRDSBundleLayout() where the size check will
fail}
\usage{
appendRDSBundle(bundle_file, objects)
}
\arguments{
\item{bundle_file}{the existing bundle file to append to}

\item{objects}{a list or an environment of objects to append}
}
\description{
Append objects to a RDSBundle file. If the first objects size is less than
the size of the index table of the existing file, instead of overwriting the
index table it is left in place and the new objects are appended at EOF.
This can be verified by using getRDSBundleLayout() where the size check will
fail
}
