% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getConnectionFromString}
\alias{getConnectionFromString}
\title{Get a connection from a chrachter, character vector or pass on a connection}
\usage{
getConnectionFromString(
  x,
  mode = "a+b",
  con_function = gzfile,
  overwrite_protect = FALSE
)
}
\arguments{
\item{x}{character or chrarcter vector to turn into a connection}

\item{mode}{to open the file(s) in}

\item{con_function}{the funciton to open the file(s) with}

\item{overwrite_protect}{bool whether to saveguard against overwriting existing files}
}
\value{
a connection or vector of connections
}
\description{
Get a connection from a chrachter, character vector or pass on a connection
}
