% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRDSBundleLayout}
\alias{getRDSBundleLayout}
\title{Returns information about the layout of the rdsb file.
Note that a failed size check usually just means that at some point an append with a new
object smaller than the current index_table was performed. In this case the low level write function
skips to EOF for writing.}
\usage{
getRDSBundleLayout(bundle_file)
}
\arguments{
\item{bundle_file}{the filename of the rdsb file}
}
\description{
Returns information about the layout of the rdsb file.
Note that a failed size check usually just means that at some point an append with a new
object smaller than the current index_table was performed. In this case the low level write function
skips to EOF for writing.
}
