% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{loadRDSBundle}
\alias{loadRDSBundle}
\title{Load the whole .rdsb file
Loads the file using furrr::future_map setup plan(multicore) or plan(mulitsession) appropriately in advance
to make use of parellelism}
\usage{
loadRDSBundle(bundle_file)
}
\arguments{
\item{bundle_file}{path to the .rdsb file}
}
\value{
the objects as a list
}
\description{
Load the whole .rdsb file
Loads the file using furrr::future_map setup plan(multicore) or plan(mulitsession) appropriately in advance
to make use of parellelism
}
