% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdsBundleOptions}
\alias{rdsBundleOptions}
\title{Set rdsBundle Options}
\usage{
rdsBundleOptions(..., reset = FALSE)
}
\arguments{
\item{...}{Named options to set}

\item{reset}{resets all options to default values if TRUE}
}
\description{
Set rdsBundle Options
}
