% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_data_object}
\alias{read_data_object}
\title{Read object from bundle file in rust}
\usage{
read_data_object(file_path, offset, size)
}
\arguments{
\item{file_path}{path to the file}

\item{offset}{the offset of the object in the file}

\item{size}{the size of the object in the file}
}
\description{
Read object from bundle file in rust
}
