% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_data_object_to_R_buffer}
\alias{read_data_object_to_R_buffer}
\title{Read object from bundle file in rust directly to a buffer in R}
\usage{
read_data_object_to_R_buffer(file_path, r_buf, offset, size)
}
\arguments{
\item{file_path}{path to the file}

\item{r_buf}{the R buffer to read into must be raw(...)}

\item{offset}{the offset of the object in the file}

\item{size}{the size of the object in the file}
}
\description{
Read object from bundle file in rust directly to a buffer in R
}
