% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{readObjectFromRDSBundle}
\alias{readObjectFromRDSBundle}
\title{Read a single object from a .rdsb file}
\usage{
readObjectFromRDSBundle(bundle_file, key, index = NULL)
}
\arguments{
\item{bundle_file}{Path to the bundle_file}

\item{key}{The key of the object in the file (string)}

\item{index}{(optional) The index table as returned by readRDSBundleIndex()}
}
\value{
The object
}
\description{
Read a single object from a .rdsb file
}
