% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{readRDSBundleIndex}
\alias{readRDSBundleIndex}
\title{Read the Index of a .rdsb File}
\usage{
readRDSBundleIndex(bundle_file)
}
\arguments{
\item{bundle_file}{A connection to the .rdsb File}
}
\value{
The index table
}
\description{
Read the Index of a .rdsb File
}
