% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce2flex}
\alias{reduce2flex}
\title{purrr::reduce2 wrapper to handle lists or environments}
\usage{
reduce2flex(.x, .y, .f, ...)
}
\arguments{
\item{.x}{a list}

\item{.y}{must be a list or an environment}

\item{.f}{the function to be reduced over .x and .y}

\item{...}{passed on to reduce2}
}
\value{
reduce2(.x, .y, .f, ...)
}
\description{
purrr::reduce2 wrapper to handle lists or environments
}
