% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{saveObjectToRDSBundle}
\alias{saveObjectToRDSBundle}
\title{Save an object to a .rdsb File}
\usage{
saveObjectToRDSBundle(bundle_file, object, name, index, current_offset)
}
\arguments{
\item{bundle_file}{The path to the bundle_file}

\item{object}{The object to save}

\item{name}{The name under which to store the object}

\item{index}{The index table to upadate}

\item{current_offset}{The current offset in the index}
}
\value{
A list containing the current index table and the current_offset
}
\description{
Save an object to a .rdsb File
}
