% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{saveRDSBundle}
\alias{saveRDSBundle}
\title{Save a list or an environment of objects to a .rdsb file}
\usage{
saveRDSBundle(bundle_file, objects)
}
\arguments{
\item{bundle_file}{filename or connection, if given a string will not overwrite,
however if given a connection is happy to overwrite}

\item{objects}{either a list or an environment of objects to save to .rdsb file}
}
\description{
Save a list or an environment of objects to a .rdsb file
}
