% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{saveRDSBundleIndex}
\alias{saveRDSBundleIndex}
\title{Save the index to the .rdbs file.}
\usage{
saveRDSBundleIndex(bundle_file, index)
}
\arguments{
\item{bundle_file}{The bundle_file connection}

\item{index}{The index table}
}
\description{
Save the index to the .rdbs file.
}
