% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{write_data_object}
\alias{write_data_object}
\title{Write the compressed data
Write object to bundle file in rust}
\usage{
write_data_object(file_path, ser_obj, offset, index_size)
}
\arguments{
\item{file_path}{path to the file}

\item{ser_obj}{serialized object}

\item{offset}{the offset of the object in the file}

\item{index_size}{the size of the index table; -1 if not applicable, only relevant for appending}
}
\description{
Write the compressed data
Write object to bundle file in rust
}
