% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeObjectsToRDSBundle}
\alias{writeObjectsToRDSBundle}
\title{Low level function to write lists of objects to .rdsb
not userfacing.}
\usage{
writeObjectsToRDSBundle(
  objects,
  bundle_con,
  index,
  current_offset,
  index_size = 0,
  names_obj = NA
)
}
\arguments{
\item{objects}{list or environment of objects}

\item{bundle_con}{connction to write to}

\item{index}{the index table}

\item{current_offset}{where the read pointer is at. DOES NOT SEEK TO THIS LOCATION}

\item{index_size}{(optional) size of the index table before writing. Must be passed if
appending to file otherwise correct alignment can not be guaranteed}

\item{names_obj}{(optional) if missing usees the names of the objects list/env}
}
\value{
the final accumulated index and current_offset in a named list
}
\description{
Low level function to write lists of objects to .rdsb
not userfacing.
}
