% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{readRDSBundleIndex}
\alias{readRDSBundleIndex}
\title{Read the Index of a .rdsb Filetesting}
\usage{
readRDSBundleIndex(bundle_file, keep_open = FALSE)
}
\arguments{
\item{bundle_file}{A connection to the .rdsb File}

\item{keep_open}{keep connection alive ?}
}
\value{
The index table
}
\description{
Read the Index of a .rdsb Filetesting
}
