% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{saveRDSBundle}
\alias{saveRDSBundle}
\title{Save a list or an environment of objects to a .rdsb file}
\usage{
saveRDSBundle(objects, bundle_file)
}
\arguments{
\item{objects}{either a list or an environment of objects to save to .rdsb file}

\item{bundle_file}{filename where to write the bundle_file}
}
\description{
Save a list or an environment of objects to a .rdsb file
}
