% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsb.R
\name{saveRDSBundleIndex}
\alias{saveRDSBundleIndex}
\title{Save the index to the .rdbs file}
\usage{
saveRDSBundleIndex(index, bundle_file)
}
\arguments{
\item{index}{The index table}

\item{bundle_file}{The bundle_file connection}
}
\description{
Save the index to the .rdbs file
}
